apps/api/pyproject.toml
[project]
name = "analyst-platform-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "claude-agent-sdk>=0.1.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "selenium>=4.40.0",
    "webdriver-manager>=4.0.2",
    "pandas>=3.0.0",
    "undetected-chromedriver>=3.5.5",
    "nodriver>=0.48.1",
]
